rename units and osparc_client and moved test_docs inside
