+++ clients/python/test/units/osparc_client/test_meta.py
@@ -1,14 +1,13 @@
 # coding: utf-8
 
 """
-    osparc.io web API
+osparc.io web API
 
-    osparc-simcore public web API specifications  # noqa: E501
+osparc-simcore public web API specifications  # noqa: E501
 
-    The version of the OpenAPI document: 0.4.0
-    Generated by: https://openapi-generator.tech
+The version of the OpenAPI document: 0.4.0
+Generated by: https://openapi-generator.tech
 """
-
 
 from __future__ import absolute_import
