+++ clients/python/test/units/osparc_client/test_task_states.py
@@ -1,24 +1,23 @@
 # coding: utf-8
 
 """
-    osparc.io web API
+osparc.io web API
 
-    osparc-simcore public web API specifications  # noqa: E501
+osparc-simcore public web API specifications  # noqa: E501
 
-    The version of the OpenAPI document: 0.4.0
-    Generated by: https://openapi-generator.tech
+The version of the OpenAPI document: 0.4.0
+Generated by: https://openapi-generator.tech
 """
-
 
 from __future__ import absolute_import
 
 import unittest
 
-from osparc import UserRoleEnum  # noqa: E501
+from osparc import TaskStates  # noqa: E501
 
 
-class TestUserRoleEnum(unittest.TestCase):
-    """UserRoleEnum unit test stubs"""
+class TestTaskStates(unittest.TestCase):
+    """TaskStates unit test stubs"""
 
     def setUp(self):
         pass
@@ -27,18 +26,18 @@ class TestUserRoleEnum(unittest.TestCase):
         pass
 
     def make_instance(self, include_optional):
-        """Test UserRoleEnum
+        """Test TaskStates
         include_option is a boolean, when False only required
         params are included, when True both required and
         optional params are included"""
-        # model = osparc.models.user_role_enum.UserRoleEnum()  # noqa: E501
+        # model = osparc.models.task_states.TaskStates()  # noqa: E501
         if include_optional:
-            return UserRoleEnum()
+            return TaskStates()
         else:
-            return UserRoleEnum()
+            return TaskStates()
 
-    def testUserRoleEnum(self):
-        """Test UserRoleEnum"""
+    def testTaskStates(self):
+        """Test TaskStates"""
         self.make_instance(include_optional=False)
         self.make_instance(include_optional=True)
